[TASK] Refactor the form.validatedControlGroup view helper
We use the TagBuilder now for rendering the control group tag
and we allow all universal tag attributes.

The tag name can now also be set with a view helper attribute.

All arguments for the view helper are moved from the render()
method to registerArgument() calls in the initializeArguments()
method.